+++ tests/Support/IntegerEnum.php
@@ -2,7 +2,7 @@
 
 declare(strict_types=1);
 
-namespace Vjik\CycleTypecast\Tests\TestEnvironments\Php81\Support;
+namespace Vjik\CycleTypecast\Tests\Support;
 
 enum IntegerEnum: int
 {