+++ src/TypeInterface.php
@@ -6,7 +6,7 @@ namespace Vjik\CycleTypecast;
 
 interface TypeInterface
 {
-    public function convertToDatabaseValue(mixed $value): mixed;
+    public function convertToDatabaseValue(mixed $value, UncastContext $context): mixed;
 
-    public function convertToPhpValue(mixed $value): mixed;
+    public function convertToPhpValue(mixed $value, CastContext $context): mixed;
 }