+++ tests/DateTimeImmutable/DateTimeImmutableToIntegerTypeTest.php
@@ -9,6 +9,7 @@ use InvalidArgumentException;
 use PHPUnit\Framework\TestCase;
 use stdClass;
 use Vjik\CycleTypecast\DateTimeImmutable\DateTimeImmutableToIntegerType;
+use Vjik\CycleTypecast\Tests\Support\ContextFactory;
 
 final class DateTimeImmutableToIntegerTypeTest extends TestCase
 {
@@ -21,30 +22,31 @@ final class DateTimeImmutableToIntegerTypeTest extends TestCase
 
         $type = new DateTimeImmutableToIntegerType();
 
-        $value = $type->convertToDatabaseValue($entityValue);
+        $value = $type->convertToDatabaseValue($entityValue, ContextFactory::uncastContext());
         $this->assertSame($databaseValue, $value);
 
-        $value = $type->convertToPhpValue($databaseValue);
+        $value = $type->convertToPhpValue($databaseValue, ContextFactory::castContext());
         $this->assertInstanceOf(DateTimeImmutable::class, $value);
         $this->assertSame($timestamp, $value->getTimestamp());
     }
 
-    public function testConertToPhpValueWithIntegerValue(): void
+    public function testConvertToPhpValueWithIntegerValue(): void
     {
         $timestamp = 1609658768;
 
         $type = new DateTimeImmutableToIntegerType();
 
-        $value = $type->convertToPhpValue($timestamp);
+        $value = $type->convertToPhpValue($timestamp, ContextFactory::castContext());
         $this->assertInstanceOf(DateTimeImmutable::class, $value);
         $this->assertSame($timestamp, $value->getTimestamp());
     }
 
-    public function testConertToPhpValueWithIncorrectValue(): void
+    public function testConvertToPhpValueWithIncorrectValue(): void
     {
         $type = new DateTimeImmutableToIntegerType();
+        $context = ContextFactory::castContext();
 
         $this->expectException(InvalidArgumentException::class);
-        $type->convertToPhpValue(new stdClass());
+        $type->convertToPhpValue(new stdClass(), $context);
     }
 }